travis: update config to make nosetests do the right thing.
When no args are given, nose evicts other Kegbot-namespace packages, in
particular kegbot.util which is necessary in several libraries/tests.

When --first-package-wins is given, nose finds zero tests.  When
--traverse-namspace is also given, all tests are discovered -- including
tests we do not want to run.

There might be some PYTHONPATH trick to ensure kegbot.pycore is the first
kegbot namespace space, thus avoiding need for --traverse-namespace and the
-e exclude regex, but I cannot find it.